requirements.txt
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
transformers==4.36.2
torch==2.1.2
together==0.2.5
python-dotenv==1.0.0
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0
accelerate==0.25.0
scikit-learn==1.3.2 